+++ src/main/java/atwoz/atwoz/admin/command/application/admin/AdminAuthService.java
@@ -1,12 +1,10 @@
 package atwoz.atwoz.admin.command.application.admin;
 
-import atwoz.atwoz.admin.command.application.admin.dto.AdminLoginRequest;
-import atwoz.atwoz.admin.command.application.admin.dto.AdminLoginResponse;
-import atwoz.atwoz.admin.command.application.admin.dto.AdminSignupRequest;
-import atwoz.atwoz.admin.command.application.admin.dto.AdminSignupResponse;
-import atwoz.atwoz.admin.command.application.admin.exception.AdminNotFoundException;
-import atwoz.atwoz.admin.command.application.admin.exception.DuplicateEmailException;
 import atwoz.atwoz.admin.command.domain.admin.*;
+import atwoz.atwoz.admin.presentation.admin.dto.AdminLoginRequest;
+import atwoz.atwoz.admin.presentation.admin.dto.AdminLoginResponse;
+import atwoz.atwoz.admin.presentation.admin.dto.AdminSignupRequest;
+import atwoz.atwoz.admin.presentation.admin.dto.AdminSignupResponse;
 import atwoz.atwoz.auth.domain.TokenProvider;
 import atwoz.atwoz.auth.domain.TokenRepository;
 import atwoz.atwoz.common.enums.Role;
