+++ src/main/java/atwoz/atwoz/admin/command/application/admin/AdminNotFoundException.java
@@ -1,4 +1,4 @@
-package atwoz.atwoz.admin.command.application.admin.exception;
+package atwoz.atwoz.admin.command.application.admin;
 
 public class AdminNotFoundException extends RuntimeException {
     public AdminNotFoundException() {
