+++ src/main/java/atwoz/atwoz/admin/presentation/admin/dto/AdminLoginResponse.java
@@ -1,4 +1,4 @@
-package atwoz.atwoz.admin.command.application.admin.dto;
+package atwoz.atwoz.admin.presentation.admin.dto;
 
 public record AdminLoginResponse(String accessToken, String refreshToken) {
 }
