+++ src/main/java/atwoz/atwoz/admin/presentation/admin/dto/AdminSignupResponse.java
@@ -1,4 +1,4 @@
-package atwoz.atwoz.admin.command.application.admin.dto;
+package atwoz.atwoz.admin.presentation.admin.dto;
 
 public record AdminSignupResponse(long id, String email, String name, String phoneNumber) {
 }
